Create README.md
This project was developed entirely in Tableau and the presentation was intended using the presentation tool in Tableau and starting in the Story sheet P&L-Q1. Since a I don't know which app the audience have I kept it simple to follow.
The file contains each worksheet with all the measures used.